agent-api/pyproject.toml
[project]
name = "agent-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "inject>=5.3.0",
    "injector>=0.24.0",
    "jieba>=0.42.1",
    "langchain>=1.2.10",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "langchain-unstructured>=1.0.1",
    "langchain-weaviate>=0.0.6",
    "markdown>=3.10.2",
    "openai>=2.21.0",
    "psycopg>=3.3.2",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "unstructured[pptx]>=0.20.8",
]
